feat: finish AnnouncementModal component
